+++ src/Tasks/Monitor/BackupDestinationStatusFactory.php
@@ -12,7 +12,8 @@ class BackupDestinationStatusFactory
         return collect($monitorConfiguration)->flatMap(function (array $monitorProperties) {
             return BackupDestinationStatusFactory::createForSingleMonitor($monitorProperties);
         })->sortBy(function (BackupDestinationStatus $backupDestinationStatus) {
-            return "{$backupDestinationStatus->backupName()}-{$backupDestinationStatus->diskName()}";
+            return $backupDestinationStatus->backupDestination()->backupName().'-'.
+                $backupDestinationStatus->backupDestination()->diskName();
         });
     }
 
@@ -21,9 +22,30 @@ class BackupDestinationStatusFactory
         return collect($monitorConfig['disks'])->map(function ($diskName) use ($monitorConfig) {
             $backupDestination = BackupDestination::create($diskName, $monitorConfig['name']);
 
-            return (new BackupDestinationStatus($backupDestination, $diskName))
-                ->setMaximumAgeOfNewestBackupInDays($monitorConfig['newestBackupsShouldNotBeOlderThanDays'])
-                ->setMaximumStorageUsageInMegabytes($monitorConfig['storageUsedMayNotBeHigherThanMegabytes']);
+            return new BackupDestinationStatus($backupDestination, static::buildHealthChecks($monitorConfig));
         });
+    }
+
+    protected static function buildHealthChecks($monitorConfig)
+    {
+        return collect(array_get($monitorConfig, 'health_checks'))->map(function ($options, $class) {
+            if (is_int($class)) {
+                $class = $options;
+                $options = [];
+            }
+
+            return static::buildHealthCheck($class, $options);
+        })->toArray();
+    }
+
+    protected static function buildHealthCheck($class, $options)
+    {
+        // A single value was passed - we'll instantiate it manually assuming it's the first argument
+        if (! is_array($options)) {
+            return new $class($options);
+        }
+
+        // A config array was given. Use reflection to match arguments
+        return app()->makeWith($class, $options);
     }
 }