+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 All notable changes to `laravel-backup` will be documented in this file.
 
+## 6.0.0 - 2018-11-28
+
+- all config keys are now snake_cased
+- health checks have been converted to their own class
+
 ## 5.12.0 - 2018-11-28
 
 - added the ability to use `backup_options` on a disk defined in the `filesystems` config file